app/server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
boto3==1.34.0
python-multipart==0.0.6
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
openai==1.3.0
anthropic==0.7.0
playwright==1.40.0
icalendar==5.0.11
weasyprint==60.2
python-dotenv==1.0.0
structlog==23.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
